+++ elisp/setup-screen-recording.el
@@ -1,0 +1,2 @@
+(set-face-attribute 'default nil :height 190)
+ (pjmai-global-mode 1)